MAINT: Move events in to events.py module
Moving event logic out of main.py into a standalone module. This improves the readability of the main module.
